image textures

+++ src/main/java/org/pt3k/materials/Checker.java
@@ -12,7 +12,9 @@ public class Checker implements Texture {
     }
 
     @Override
-    public Vec3 value(float u, double v, Vec3 p) {
+    public Vec3 value(float u, float v, Vec3 p) {
+        //System.out.println(u + " " + v);
+
         double sines = Math.sin(10*p.getX())*Math.sin(10*p.getY())
                 *Math.sin(10*p.getZ());
         if(sines < 0) {
